+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+- Improved slurping and barfing (still only HTML)
+- New edit command: transpose siblings
+- New edit command: splice
+
 ## [0.2.0] - 2021-02-26
 
 -  First public release